chore: update to version 0.15.0
