+++ pubspec.yaml
@@ -15,7 +15,7 @@ description: A Flutter Add-to-app Module for the Reclaim InApp Verifier SDK
 # This version is used _only_ for the Runner app, which is used if you just do
 # a `flutter run` or a `flutter make-host-app-editable`. It has no impact
 # on any other native host app that you embed your Flutter project into.
-version: 0.12.0
+version: 0.15.0
 
 environment:
   sdk: '>=3.8.0 <4.0.0'
