+++ scripts/inapp/reactnative.sh
@@ -15,7 +15,7 @@ export RN_CLONE_DIR="$WORK_DIR/build/reactnative-sdk";
 if [[ -z "$PACKAGE_CLONE_USER" ]]; then
     git clone git@github.com:reclaimprotocol/reclaim-inapp-reactnative-sdk.git $RN_CLONE_DIR;
 else
-    git clone https://$PACKAGE_CLONE_USER:$PACkAGE_CLONE_PASSWD@github.com/reclaimprotocol/reclaim-inapp-reactnative-sdk.git $RN_CLONE_DIR;
+    git clone https://$PACKAGE_CLONE_USER:$PACKAGE_CLONE_PASSWD@github.com/reclaimprotocol/reclaim-inapp-reactnative-sdk.git $RN_CLONE_DIR;
 fi
 
 cd $RN_CLONE_DIR;
